backend/pyproject.toml
[project]
name = "rag-voice-backend"
version = "0.1.0"
description = "CPU-only local RAG demo with voice interface"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "faster-whisper>=1.0.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",
    "chonkie[neural]>=1.0.0",
    "accelerate>=0.25.0",
]

[project.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.26.0",
]
